+++ CHANGELOG.md
@@ -9,8 +9,10 @@
 
 ### Changed
 
-- **breaking change**: rename `image-generation` event type to `generate-image`
 - **breaking change**: rename `embedding` event type to `embed`
+- **breaking change**: rename `image-generation` event type to `generate-image`
+- **breaking change**: rename `speech-generation` event type to `generate-speech`
+- **breaking change**: rename `speech-streaming` event type to `stream-speech`
 
 ## v0.49.0 - 2023-10-21